# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Golf Charity Subscription Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scores:
    post:
      operationId: addScore
      tags: [scores]
      summary: Add a new golf score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddScoreBody"
      responses:
        "201":
          description: Score added (rolling 5 maintained)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
  /scores/{scoreId}:
    patch:
      operationId: updateScore
      tags: [scores]
      summary: Update a golf score
      parameters:
        - name: scoreId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScoreBody"
      responses:
        "200":
          description: Score updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
    delete:
      operationId: deleteScore
      tags: [scores]
      summary: Delete a golf score
      parameters:
        - name: scoreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /scores/user/{userId}:
    get:
      operationId: getUserScores
      tags: [scores]
      summary: Get a specific user's scores (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"

  # Charities
...
  /draws:
    get:
      operationId: listDraws
      tags: [draws]
      summary: List all draws
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, simulated, published]
      responses:
        "200":
          description: Draws list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draw"
    post:
      operationId: createDraw
      tags: [draws]
      summary: Create a new draw (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrawBody"
      responses:
        "201":
          description: Draw created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
...
  /draws/{drawId}:
    get:
      operationId: getDraw
      tags: [draws]
      summary: Get a draw by ID
      parameters:
        - name: drawId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
  /draws/{drawId}/simulate:
    post:
      operationId: simulateDraw
      tags: [draws]
      summary: Run draw simulation (admin)
      parameters:
        - name: drawId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Simulation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawSimulationResult"
  /draws/{drawId}/publish:
    post:
      operationId: publishDraw
      tags: [draws]
      summary: Publish draw results (admin)
      parameters:
        - name: drawId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"

  # Winners
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Score:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        value:
          type: integer
          minimum: 1
          maximum: 45
        playedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - value
        - playedAt
        - createdAt
    AddScoreBody:
      type: object
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 45
        playedAt:
          type: string
          format: date-time
      required:
        - value
        - playedAt
    UpdateScoreBody:
      type: object
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 45
        playedAt:
          type: string
          format: date-time

    # Charity
...
    Draw:
      type: object
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        status:
          type: string
          enum: [pending, simulated, published]
        drawType:
          type: string
          enum: [random, algorithmic]
        drawnNumbers:
          type: array
          items:
            type: integer
          nullable: true
        totalPrizePool:
          type: number
        jackpotAmount:
          type: number
        fourMatchAmount:
          type: number
        threeMatchAmount:
          type: number
        jackpotRolledOver:
          type: boolean
        participantCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - month
        - year
        - status
        - drawType
        - totalPrizePool
        - jackpotAmount
        - fourMatchAmount
        - threeMatchAmount
        - jackpotRolledOver
        - participantCount
        - createdAt
    CreateDrawBody:
      type: object
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        drawType:
          type: string
          enum: [random, algorithmic]
          default: random
      required:
        - month
        - year
    DrawSimulationResult:
      type: object
      properties:
        draw:
          $ref: "#/components/schemas/Draw"
        drawnNumbers:
          type: array
          items:
            type: integer
        fiveMatchWinners:
          type: array
          items:
            $ref: "#/components/schemas/DrawWinnerPreview"
        fourMatchWinners:
          type: array
          items:
            $ref: "#/components/schemas/DrawWinnerPreview"
        threeMatchWinners:
          type: array
          items:
            $ref: "#/components/schemas/DrawWinnerPreview"
      required:
        - draw
        - drawnNumbers
        - fiveMatchWinners
        - fourMatchWinners
        - threeMatchWinners
    DrawWinnerPreview:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        matchedNumbers:
          type: array
          items:
            type: integer
        matchCount:
          type: integer
        prizeAmount:
          type: number
      required:
        - userId
        - userName
        - matchedNumbers
        - matchCount
        - prizeAmount

    # Winners
...