# lib/api-zod/src/generated/types/draw.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DrawDrawType } from "./drawDrawType";
import type { DrawStatus } from "./drawStatus";

export interface Draw {
  id: number;
  month: number;
  year: number;
  status: DrawStatus;
  drawType: DrawDrawType;
  drawnNumbers?: number[] | null;
  totalPrizePool: number;
  jackpotAmount: number;
  fourMatchAmount: number;
  threeMatchAmount: number;
  jackpotRolledOver: boolean;
  participantCount: number;
  publishedAt?: Date | null;
  createdAt: Date;
}