# lib/api-zod/src/generated/types/drawSimulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Draw } from "./draw";
import type { DrawWinnerPreview } from "./drawWinnerPreview";

export interface DrawSimulationResult {
  draw: Draw;
  drawnNumbers: number[];
  fiveMatchWinners: DrawWinnerPreview[];
  fourMatchWinners: DrawWinnerPreview[];
  threeMatchWinners: DrawWinnerPreview[];
}